requirements.txt
# 基础数据处理库
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
 Pillow>=8.3.0  # 图像处理

# 深度学习框架
torch>=1.10.0
torchvision>=0.11.0

# 工具库
chardet>=4.0.0  # 编码检测
tqdm>=4.62.0    # 进度条显示
json5>=0.9.6    # JSON处理增强（可选）

# 可选：CUDA支持（根据实际环境调整版本）
# torch==1.13.1+cu117
# torchvision==0.14.1+cu117